refactor: input 필드 컴포넌트 위치 수정
