+++ src/components/auth/Form/AuthForm/index.tsx
@@ -1,5 +1,5 @@
 import PrimaryButton from '@/components/@common/Button/PrimaryButton';
-import InputField from '@/components/auth/InputField';
+import InputField from '@/components/@common/Form/InputField';
 import Timer from '@/components/auth/Timer';
 import { AUTH_ERROR_MSG } from '@/constants/message';
 import { AUTH_PATTERN } from '@/constants/pattern';
